# openapi.yaml
openapi: 3.1.0
info:
  title: GPT Dev Log API - Full Endpoint Access
  version: "1.5"
servers:
  - url: https://flask-inventory-api.onrender.com
paths:
  /sheet/write_row:
    post:
      operationId: writeSheetRow
      summary: Write a row to any sheet in the Inventory_Tracker spreadsheet
      description: >
        Inserts or updates a single row in the specified sheet.  
        Supports all core, warehouse, system, alert, knowledge, analytics, and media headers.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - sheet_name
              properties:
                # === CORE & UNIVERSAL FIELDS ===
                sheet_name:
                  type: string
                  description: Target sheet to write to
                unified_log_id:
                  type: string
                  description: Global UUID for each record
                origin_sheet:
                  type: string
                  description: Sheet from which the record originated
                linked_sheets:
                  type: string
                  description: Comma-separated related sheets
                linked_log_ids:
                  type: string
                  description: Related Unified Log IDs
                timestamp:
                  type: string
                  format: date-time
                category:
                  type: string
                tags:
                  type: string
                summary:
                  type: string
                details:
                  type: string
                status:
                  type: string
                  description: Record state (Open/Closed/Complete/etc.)
                priority:
                  type: string
                outcome:
                  type: string
                outcome_detail:
                  type: string
                version:
                  type: string
                environment:
                  type: string
                location:
                  type: string
                duration:
                  type: string

                # === INTEGRATION & EXTERNAL SYSTEMS ===
                zoho_id:
                  type: string
                  description: Linked Zoho Inventory or Desk record
                ticket_id:
                  type: string
                shipment_id:
                  type: string
                booking_id:
                  type: string
                customer_name:
                  type: string
                purchase_order:
                  type: string
                sales_order:
                  type: string
                carrier:
                  type: string
                pickup_date:
                  type: string
                  format: date
                delivery_date:
                  type: string
                  format: date
                rma_number:
                  type: string

                # === WAREHOUSE OPERATIONS ===
                item_name:
                  type: string
                category:
                  type: string
                serial_number:
                  type: string
                quantity:
                  type: number
                unit_cost:
                  type: number
                supplier:
                  type: string
                condition:
                  type: string
                bay_id:
                  type: string
                  description: Physical bay or location ID
                available_quantity:
                  type: number
                reserved_quantity:
                  type: number
                last_updated:
                  type: string
                  format: date-time
                assembled_by:
                  type: string
                assembly_stage:
                  type: string
                task:
                  type: string
                assigned_to:
                  type: string
                start_time:
                  type: string
                  format: date-time
                end_time:
                  type: string
                  format: date-time
                components:
                  type: string
                resolution_log_id:
                  type: string
                po_number:
                  type: string
                so_number:
                  type: string

                # === ALERTS & SYSTEM EVENTS ===
                alert_type:
                  type: string
                severity:
                  type: string
                triggered_by:
                  type: string
                trigger_phrase:
                  type: string
                trigger_time:
                  type: string
                  format: date-time
                action:
                  type: string
                action_result:
                  type: string
                confirmation_status:
                  type: string
                resolution_notes:
                  type: string
                message:
                  type: string
                context:
                  type: string
                system_event:
                  type: string
                integration_name:
                  type: string
                endpoint:
                  type: string
                request_payload:
                  type: string
                response_code:
                  type: string
                response_message:
                  type: string
                result:
                  type: string
                error_code:
                  type: string

                # === KNOWLEDGE / TASKS / IMPROVEMENTS ===
                title:
                  type: string
                notes:
                  type: string
                suggestion:
                  type: string
                impact_level:
                  type: string
                due_date:
                  type: string
                  format: date
                guide_title:
                  type: string
                steps:
                  type: string
                test_case:
                  type: string
                scenario:
                  type: string
                expected_behavior:
                  type: string
                actual_behavior:
                  type: string
                validated_output:
                  type: string
                expected:
                  type: string
                validator:
                  type: string
                comments:
                  type: string

                # === ANALYTICS / DASHBOARD METRICS ===
                log_type:
                  type: string
                logs_created:
                  type: number
                errors_logged:
                  type: number
                tasks_completed:
                  type: number
                open_issues:
                  type: number
                shipments_completed:
                  type: number
                returns_processed:
                  type: number
                average_turnaround_time:
                  type: string
                purchase_order_count:
                  type: number
                sales_order_count:
                  type: number
                highlights:
                  type: string
                next_actions:
                  type: string
                linked_logs:
                  type: string

                # === MEDIA / FILE RECORDS ===
                media_type:
                  type: string
                  description: Type of file (photo, screenshot, scan)
                file_name:
                  type: string
                media_url:
                  type: string
                description:
                  type: string
                captured_text:
                  type: string
                  description: Text extracted from uploaded image via OCR
                reference_images:
                  type: string
                  description: One or more image URLs (comma-separated if multiple)
      security:
        - BearerAuth: []

  /sheet/write_passthrough_log:
    post:
      operationId: writePassthroughLog
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                sheet_name:
                  type: string
                log_id:
                  type: string
                status:
                  type: string
                notes:
                  type: string
                timestamp:
                  type: string
                summary:
                  type: string
                action:
                  type: string
      security:
        - BearerAuth: []

  /upload/base64:
    post:
      operationId: uploadBase64Screenshot
      summary: Uploads a base64-encoded image to Google Drive
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                base64_data:
                  type: string
                  description: The base64-encoded image content
                filename:
                  type: string
                  description: Optional filename (e.g. screenshot-20251009.png)
                folder_id:
                  type: string
                  description: Google Drive folder ID for upload
      responses:
        '200':
          description: Upload successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  url:
                    type: string
                    description: Public Drive link to the uploaded file
        '400':
          description: Bad request
        '500':
          description: Server error
      security:
        - BearerAuth: []

  /upload/screenshot:
    post:
      operationId: uploadScreenshot
      summary: Uploads a screenshot to Google Drive and returns a public URL
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                screenshot:
                  type: string
                  format: binary
                  description: The image file to upload
      responses:
        '200':
          description: Upload successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  url:
                    type: string
                    format: uri
                    description: Link to the uploaded screenshot on Google Drive
        '500':
          description: Upload failed due to server error

  /sheet/get_item:
    post:
      operationId: getInventoryItemBySheet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                sheet_name:
                  type: string
                item_name:
                  type: string
      security:
        - BearerAuth: []

  /sheet/get_headers:
    post:
      operationId: getSheetHeaders
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                sheet_name:
                  type: string
      security:
        - BearerAuth: []

  /sheet/get_all:
    post:
      operationId: getStructuredSheet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                sheet_name:
                  type: string
      security:
        - BearerAuth: []

  /sheet/get_by_location:
    post:
      operationId: getByLocationBySheet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                sheet_name:
                  type: string
                location_id:
                  type: string
      security:
        - BearerAuth: []

  /updateSheetHeaders:
    post:
      operationId: updateSheetHeaders
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                sheet_name:
                  type: string
                headers:
                  type: array
                  items:
                    type: string
      security:
        - BearerAuth: []

  /log:
    post:
      operationId: logEvent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                timestamp:
                  type: string
                source:
                  type: string
                severity:
                  type: string
                message:
                  type: string
                details:
                  type: string
      security:
        - BearerAuth: []

  /habits/update:
    post:
      operationId: updateHabit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                date:
                  type: string
                habit:
                  type: string
                status:
                  type: string
                notes:
                  type: string
      security:
        - BearerAuth: []

  /goals/update:
    post:
      operationId: updateGoal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                goal_name:
                  type: string
                target_date:
                  type: string
                progress:
                  type: string
                status:
                  type: string
                priority:
                  type: string
      security:
        - BearerAuth: []

  /mental_state/update:
    post:
      operationId: updateMentalState
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                date:
                  type: string
                mood:
                  type: string
                energy:
                  type: string
                notes:
                  type: string
      security:
        - BearerAuth: []

  /submitJournalV3:
    post:
      operationId: submitJournalEntry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                sheet_name:
                  type: string
                entry_data:
                  type: object
                  additionalProperties: true
                write_key:
                  type: string
      security:
        - BearerAuth: []

  /sheet/create:
    post:
      operationId: createSheet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - sheet_name
              properties:
                sheet_name:
                  type: string
                headers:
                  type: array
                  items:
                    type: string
      security:
        - BearerAuth: []

  /sheet/rename:
    post:
      operationId: renameSheet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - old_name
                - new_name
              properties:
                old_name:
                  type: string
                new_name:
                  type: string
      security:
        - BearerAuth: []

  /sheet/delete:
    post:
      operationId: deleteSheet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - sheet_name
              properties:
                sheet_name:
                  type: string
      security:
        - BearerAuth: []

  /sheet/list_all:
    get:
      operationId: listAllSheets
      responses:
        "200":
          description: A list of sheet names
      security:
        - BearerAuth: []

securitySchemes:
  BearerAuth:
    type: http
    scheme: bearer